KINO-68 extracted methods regarding TMDB from MovieService to TMDBService
