+++ src/main/java/dat3/kino/controllers/MovieController.java
@@ -2,9 +2,8 @@ package dat3.kino.controllers;
 
 import dat3.kino.dto.response.MovieDetailsResponse;
 import dat3.kino.dto.response.MovieResponse;
-import dat3.kino.dto.response.TmdbMovieDetailsResponse;
-import dat3.kino.entities.Movie;
 import dat3.kino.services.MovieService;
+import dat3.kino.services.TMDBService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -15,9 +14,11 @@ import java.util.List;
 public class MovieController {
 
     private final MovieService movieService;
+    private final TMDBService tmdbService;
 
-    public MovieController(MovieService movieService) {
+    public MovieController(MovieService movieService, TMDBService tmdbService) {
         this.movieService = movieService;
+        this.tmdbService = tmdbService;
     }
 
     @GetMapping("/movies")
@@ -27,7 +28,7 @@ public class MovieController {
 
     @GetMapping("/movies/TMDB/{id}")
     public ResponseEntity<MovieDetailsResponse> getTMDBMovie(@PathVariable Integer id) {
-        return ResponseEntity.ok(movieService.findMovieFromTMDB(id));
+        return ResponseEntity.ok(tmdbService.findMovieFromTMDB(id));
     }
